+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+LTS Breaking: Removed the following previously deprecated items: [#1981](https://github.com/zowe/zowe-cli/pull/1981)
+  - Moved the many constants from `zowe-cli/packages/cli/src/Constants.ts` to `zowe-cli/packages/core/src/constants/Core.constants.ts`
+  - Removing `ZosFilesCreateExtraOptions.showAttributes` without replacement
+  - Moved all constants from `zowe-cli/packages/cli/src/zostso/constants/ZosTso.constants.ts` to  `@zowe/zos-tso-for-zowe-sdk`
+  - Removed `isStderrEmptyForProfilesCommand` use `stripProfileDeprecationMessages` from `zowe-cli/__tests__/__packages__/cli-test-utils/src/TestUtils.ts` instead
+  - Removed  `allDataSetsArchived`, `datasetsDownloadedSuccessfully`, `noDataSetsMatchingPatternRemain` and `onlyEmptyPartitionedDataSets` from    ZosFiles.messages.ts
+  - Removed `getSpoolDownloadFile` use `getSpoolDownloadFilePath` instead
+  - Removed constants from ZosmfSession
+    - ZOSMF_OPTION_HOST_PROFILE use ZOSMF_OPTION_HOST instead
+    - ZOSMF_OPTION_USER_PROFILE use ZOSMF_OPTION_USER instead
+    - ZOSMF_OPTION_PASSWORD_PROFILE use ZOSMF_OPTION_PASSWORD instead
+  - Removed constants from SshSession.ts
+    - SSH_OPTION_USER_PROFILE use SSH_OPTION_USER
+    - SSH_OPTION_HOST_PROFILE use SSH_OPTION_HOST
+  - Removed zosmfProfile from `ZosFilesBase.handler.ts`
+  - Removed statCmdFlag as an export from Shell.ts
+
 ## `8.0.0-next.202401262128`
 
 - Enhancement: Adding `--binary` and `--encoding` options to `zosfiles edit`